// src/inscripciones/dto/updateInscripciones.dto.ts
export class UpdateInscripcionesDto {
    idcategoria?: number
    idcorredor?: number
    placa?: number
    fecha_hora?: Date
    estado?: number
    pagado?: number
    pago_fecha_hora?: Date
    horalargada?: string
    observaciones?: string
    proline?: string
    cc9?: string
    idprecio?: number
    placa_pedida?: number
    placa_pedida_fecha?: Date
    etiqueta_exp?: number
    etiqueta_exp_fecha?: Date
    idmercadopago?: number
    facturado?: number
    idequipo?: number
}